# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/farmer/farm": {
      "post": {
        "operationId": "FarmerController_createFarm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFarmDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Farms retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CreateFarmResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Farmer"
        ]
      },
      "get": {
        "operationId": "FarmerController_findFarms",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Filter farms by name (case-insensitive)",
            "schema": {
              "example": "Green Valley Farm",
              "type": "string"
            }
          },
          {
            "name": "location",
            "required": false,
            "in": "query",
            "description": "Match address, district, or state (case-insensitive partial match)",
            "schema": {
              "example": "Selangor",
              "type": "string"
            }
          },
          {
            "name": "minSize",
            "required": false,
            "in": "query",
            "description": "Minimum farm size",
            "schema": {
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "maxSize",
            "required": false,
            "in": "query",
            "description": "Maximum farm size",
            "schema": {
              "example": 100,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by verification status",
            "schema": {
              "example": "VERIFIED",
              "type": "string",
              "enum": [
                "PENDING",
                "VERIFIED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by produce category the farm grows",
            "schema": {
              "example": "Vegetable",
              "type": "string"
            }
          },
          {
            "name": "sizeUnit",
            "required": false,
            "in": "query",
            "description": "Unit for the provided size filters",
            "schema": {
              "example": "HECTARE",
              "type": "string",
              "enum": [
                "HECTARE",
                "ACRE",
                "SQUARE_METER"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Farms retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/FarmListRespondDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Farmer"
        ]
      }
    },
    "/farmer/farm/{farmId}": {
      "get": {
        "operationId": "FarmerController_findFarm",
        "parameters": [
          {
            "name": "farmId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by produce name or batch ID (case-insensitive)",
            "schema": {
              "example": "mango",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter produce by status",
            "schema": {
              "example": "ONCHAIN_CONFIRMED",
              "type": "string",
              "enum": [
                "DRAFT",
                "PENDING_CHAIN",
                "ONCHAIN_CONFIRMED",
                "IN_TRANSIT",
                "ARRIVED",
                "RETAILER_VERIFIED",
                "ARCHIVED"
              ]
            }
          },
          {
            "name": "harvestFrom",
            "required": false,
            "in": "query",
            "description": "Filter produce harvested on or after this date (ISO string)",
            "schema": {
              "example": "2024-01-01",
              "type": "string"
            }
          },
          {
            "name": "harvestTo",
            "required": false,
            "in": "query",
            "description": "Filter produce harvested on or before this date (ISO string)",
            "schema": {
              "example": "2024-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Farm retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/FarmDetailResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Farmer"
        ]
      },
      "patch": {
        "operationId": "FarmerController_updateFarm",
        "parameters": [
          {
            "name": "farmId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFarmDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Farmer"
        ]
      },
      "delete": {
        "operationId": "FarmerController_deleteFarm",
        "parameters": [
          {
            "name": "farmId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Farmer"
        ]
      }
    },
    "/farmer/subsidy": {
      "post": {
        "operationId": "FarmerController_createSubsidy",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestSubsidyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Farmer"
        ]
      },
      "get": {
        "operationId": "FarmerController_findSubsidies",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "programName",
            "required": false,
            "in": "query",
            "description": "Filter by program name (case-insensitive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appliedDateFrom",
            "required": false,
            "in": "query",
            "description": "Filter subsidies applied on or after this date (ISO string)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appliedDateTo",
            "required": false,
            "in": "query",
            "description": "Filter subsidies applied on or before this date (ISO string)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amountMin",
            "required": false,
            "in": "query",
            "description": "Minimum subsidy amount",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "amountMax",
            "required": false,
            "in": "query",
            "description": "Maximum subsidy amount",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by subsidy status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "DISBURSED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subsidies retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SubsidyResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Farmer"
        ]
      }
    },
    "/farmer/farms/{farmId}/produce": {
      "post": {
        "operationId": "FarmerController_createProduce",
        "parameters": [
          {
            "name": "farmId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProduceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Produce created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CreateProduceResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Farmer"
        ]
      }
    },
    "/farmer/produce": {
      "get": {
        "operationId": "FarmerController_findProduces",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter produce batches by status",
            "schema": {
              "example": "ARRIVED",
              "type": "string",
              "enum": [
                "DRAFT",
                "PENDING_CHAIN",
                "ONCHAIN_CONFIRMED",
                "IN_TRANSIT",
                "ARRIVED",
                "RETAILER_VERIFIED",
                "ARCHIVED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by produce name, farm name, or batch ID (case-insensitive)",
            "schema": {
              "example": "mango",
              "type": "string"
            }
          },
          {
            "name": "harvestFrom",
            "required": false,
            "in": "query",
            "description": "Filter batches harvested on or after this date (ISO string)",
            "schema": {
              "example": "2024-01-01",
              "type": "string"
            }
          },
          {
            "name": "harvestTo",
            "required": false,
            "in": "query",
            "description": "Filter batches harvested on or before this date (ISO string)",
            "schema": {
              "example": "2024-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produces retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ProduceListResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Farmer"
        ]
      }
    },
...
    "/subsidy": {
      "post": {
        "operationId": "SubsidyController_requestSubsidy",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestSubsidyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subsidy requested successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SubsidyResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "subsidy"
        ]
      },
      "get": {
        "operationId": "SubsidyController_listSubsidies",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "programName",
            "required": false,
            "in": "query",
            "description": "Filter by program name (case-insensitive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appliedDateFrom",
            "required": false,
            "in": "query",
            "description": "Filter subsidies applied on or after this date (ISO string)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appliedDateTo",
            "required": false,
            "in": "query",
            "description": "Filter subsidies applied on or before this date (ISO string)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amountMin",
            "required": false,
            "in": "query",
            "description": "Minimum subsidy amount",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "amountMax",
            "required": false,
            "in": "query",
            "description": "Maximum subsidy amount",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by subsidy status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "DISBURSED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subsidies retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SubsidyResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "subsidy"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "The backend API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CommonResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "CreateFarmResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "size": {
            "type": "number",
            "description": "Farm size value"
          },
          "sizeUnit": {
            "type": "string",
            "description": "Unit for the farm size measurement",
            "enum": [
              "HECTARE",
              "ACRE",
              "SQUARE_METER"
            ],
            "example": "HECTARE"
          },
          "produceCategories": {
            "description": "Produce categories grown on the farm",
            "example": [
              "Paddy",
              "Vegetable"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "verificationStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "VERIFIED",
              "REJECTED"
            ],
            "description": "Verification status set by government agency"
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "state",
          "district",
          "size",
          "sizeUnit",
          "produceCategories",
          "verificationStatus"
        ]
      },
      "CreateFarmDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "size": {
            "type": "number",
            "description": "Farm size value"
          },
          "sizeUnit": {
            "type": "string",
            "description": "Unit for the farm size measurement",
            "enum": [
              "HECTARE",
              "ACRE",
              "SQUARE_METER"
            ],
            "example": "HECTARE"
          },
          "produceCategories": {
            "description": "Produce categories grown on the farm",
            "example": [
              "Paddy",
              "Vegetable"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "address",
          "state",
          "district",
          "size",
          "sizeUnit",
          "produceCategories"
        ]
      },
...
      "FarmListRespondDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "size": {
            "type": "number",
            "description": "Farm size value"
          },
          "sizeUnit": {
            "type": "string",
            "description": "Unit for the recorded farm size",
            "enum": [
              "HECTARE",
              "ACRE",
              "SQUARE_METER"
            ]
          },
          "produceCategories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "farmDocuments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FarmDocumentDto"
            }
          },
          "verificationStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "VERIFIED",
              "REJECTED"
            ],
            "description": "Verification status set by government agency"
          },
          "produces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProduceListResponseDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "state",
          "district",
          "size",
          "sizeUnit",
          "produceCategories",
          "farmDocuments",
          "verificationStatus",
          "produces"
        ]
      },
      "FarmDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "size": {
            "type": "number",
            "description": "Farm size value"
          },
          "sizeUnit": {
            "type": "string",
            "description": "Unit for the recorded farm size",
            "enum": [
              "HECTARE",
              "ACRE",
              "SQUARE_METER"
            ]
          },
          "produceCategories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "farmDocuments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FarmDocumentDto"
            }
          },
          "farmerId": {
            "type": "string",
            "description": "Owner of the farm"
          },
          "verificationStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "VERIFIED",
              "REJECTED"
            ],
            "description": "Verification status set by government agency"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "produces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProduceListResponseDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "state",
          "district",
          "size",
          "sizeUnit",
          "produceCategories",
          "farmDocuments",
          "farmerId",
          "verificationStatus",
          "createdAt",
          "produces"
        ]
      },
...
    }
  }
}